clock test

+++ current.hpp
@@ -4,8 +4,8 @@
 #define CURRENT_H
 
 // #define VIN A0
-const float vcc    = 5.00;// supply voltage 5V or 3.3V
-const float factor = 0.02;// 20mV/A is the factor
+const double vcc    = 5.00;// supply voltage 5V or 3.3V
+const double factor = 0.02;// 20mV/A is the factor
 // int ledPin = 3;      // LED connected to digital pin 9
 // int val = 0;         // variable to store the read value
 
@@ -23,7 +23,7 @@ namespace Sensor {
         ~CurrentSensor();
 
         void tick();
-        float read();
+        double read();
     };
 
     CurrentSensor::CurrentSensor(uint8_t pin) {
@@ -43,12 +43,12 @@ namespace Sensor {
         readings++;
     }
 
-    float CurrentSensor::read() {
+    double CurrentSensor::read() {
         //Robojax.com ACS758 Current Sensor 
-        float average = (float)totalReading / (float)readings;
-        float voltage = (5.0 / 1023.0) * (float)average;// Read the voltage from sensor
-        // float a = voltage - (vcc * 0.5) + 0.007;// 0.007 is a value to make voltage zero when there is no current
-        // float current = voltage / a;
+        double average = (double)totalReading / (double)readings;
+        double voltage = (5.0 / 1023.0) * (double)average;// Read the voltage from sensor
+        // double a = voltage - (vcc * 0.5) + 0.007;// 0.007 is a value to make voltage zero when there is no current
+        // double current = voltage / a;
 
         totalReading = 0;
         readings = 0;
